chore(connect): fix type of connect HOC
