+++ packages/types/utils.ts
@@ -18,3 +18,14 @@ export type DeepPartial<T> = T extends Array<infer U>
 
 // Omits any property found in the passed object.
 export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
+
+// Resolves the actions for things like Action and Derived.
+export type ResolveActions<Actions extends Package["state"]> = {
+  [P in keyof Actions]: Actions[P] extends (
+    ...a: any
+  ) => (arg: infer Arg) => void // Turns "state => args => {}" into "args => {}"
+    ? (arg: Arg) => void // Turns "state => {}" into "() => {}"
+    : Actions[P] extends (state: Package["state"]) => any
+    ? () => void
+    : ResolveActions<Actions[P]>
+};
